refactor: change element and recipe schema to use 32 bit integers in id

+++ src/backend/schema/element.go
@@ -1,7 +1,7 @@
 package schema
 
 type Element struct {
-	ID       int64
+	ID       int32
 	Name     string
 	ImageUrl string
 }
