+++ src/backend/schema/recipe.go
@@ -1,7 +1,7 @@
 package schema
 
 type Recipe struct {
-	ResultID      int64
-	Dependency1ID int64
-	Dependency2ID int64
+	ResultID      int32
+	Dependency1ID int32
+	Dependency2ID int32
 }
